+++ language/runtime/ceylon/language/Natural.java
@@ -93,8 +93,8 @@ public final class Natural
 
 	@TypeInfo(value="ceylon.language.Integer")
     @Override
-    public long getInteger() {
-        return value;
+    public int getInteger() {
+        return (int) value;
     }
 
 	@TypeInfo(value="ceylon.language.Float")
